Switch background and foreground color in habamax theme
Found the previous behaviour rather confusing.
You go to the opening bracket with your cursor and the matching bracket
is highlighted in the same way the cursor is usually highlighted. That
way you may think you are at the closing bracket but actually the cursor
is still on the opening bracket.