sway: fix double suspend
Previous behavior: after manually suspending, when I pressed a key on
the keyboard, the system would wake up and go right back to sleep. The
second key press would actually wake up the system.

The reason is, AFAICT, that I locked twice: the first time via
`systemctl suspend` and when the first key press woke up the system,
swayidle (or swaylock?) triggered the actual lock and another suspend,
so the system went back to sleep.